Move field-constants.ts
