+++ src/components/filter/criteria-based/criteria-based-filter-utils.ts
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import { FieldConstants } from '../constants/field-constants';
+import { FieldConstants } from '../../../utils/field-constants';
 import { FreePropertiesTypes } from './filter-properties';
 import {
     PROPERTY_NAME,
