+++ src/components/filter/criteria-based/filter-properties.tsx
@@ -20,7 +20,7 @@ import {
     VoltageLevel,
 } from '../../../utils/equipment-types';
 import { areArrayElementsUnique } from '../../../utils/functions';
-import { FieldConstants } from '../constants/field-constants';
+import { FieldConstants } from '../../../utils/field-constants';
 import yup from '../../../utils/yup-config';
 import FilterFreeProperties from './filter-free-properties';
 import {
