+++ src/components/filter/expert/expert-filter-form.tsx
@@ -25,7 +25,7 @@ import * as yup from 'yup';
 import { FieldType } from './expert-filter.type';
 import { v4 as uuid4 } from 'uuid';
 import { useIntl } from 'react-intl';
-import { FieldConstants } from '../constants/field-constants';
+import { FieldConstants } from '../../../utils/field-constants';
 import CustomReactQueryBuilder from '../../react-query-builder-inputs/custom-react-query-builder';
 import InputWithPopupConfirmation from '../../react-hook-form/select-inputs/input-with-popup-confirmation';
 import SelectInput from '../../react-hook-form/select-inputs/select-input';
