+++ src/components/filter/utils/filter-form-utils.ts
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-import { FieldConstants } from '../constants/field-constants';
+import { FieldConstants } from '../../../utils/field-constants';
 import RangeInput from '../../react-hook-form/range-input';
 import { FunctionComponent } from 'react';
 import CountriesInput from '../../react-hook-form/select-inputs/countries-input.tsx';
