+++ src/components/react-query-builder-inputs/property-value-editor.tsx
@@ -12,7 +12,7 @@ import { ValueEditorProps } from 'react-querybuilder';
 import useValid from './use-valid.ts';
 import { useIntl } from 'react-intl';
 
-import { FieldConstants } from '../filter/constants/field-constants.ts';
+import { FieldConstants } from '../../utils/field-constants';
 import { OPERATOR_OPTIONS } from '../filter/constants/expert-filter-constants.ts';
 import { usePredefinedProperties } from '../../hooks/predefined-properties-hook.ts';
