+++ src/index.js
@@ -48,7 +48,7 @@ export {
     noSelectionForCopy,
 } from './utils/equipment-types';
 
-export { FieldConstants } from './components/filter/constants/field-constants';
+export { FieldConstants } from './utils/field-constants';
 
 export {
     GRIDSUITE_DEFAULT_PRECISION,
